+++ src/app/modules/instructor/instructor.service.ts
@@ -1,3 +1,5 @@
+import status from 'http-status';
+import ApiError from '../../errors/ApiError';
 import { IInstructor } from './instructor.interface';
 import { Instructor } from './instructor.model';
 
@@ -8,6 +10,9 @@ const createInstructor = async (payload: IInstructor) => {
 
 const getInstructorById = async (id: string) => {
     const result = await Instructor.findById(id);
+    if (!result) {
+        throw new ApiError(status.NOT_FOUND, 'Instructor not found');
+    }
     return result;
 };
 
@@ -17,6 +22,10 @@ const getAllInstructors = async () => {
 };
 
 const updateInstructor = async (id: string, payload: Partial<IInstructor>) => {
+    const instructor = await Instructor.findById(id);
+    if (!instructor) {
+        throw new ApiError(status.NOT_FOUND, 'Instructor not found');
+    }
     const result = await Instructor.findByIdAndUpdate(id, payload, {
         new: true,
     });
@@ -24,6 +33,10 @@ const updateInstructor = async (id: string, payload: Partial<IInstructor>) => {
 };
 
 const deleteInstructor = async (id: string) => {
+    const instructor = await Instructor.findById(id);
+    if (!instructor) {
+        throw new ApiError(status.NOT_FOUND, 'Instructor not found');
+    }
     const result = await Instructor.findByIdAndDelete(id);
     return result;
 };
